XCropAIService/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==45.0.7
exceptiongroup==1.3.1
fastapi==0.128.2
flatbuffers==25.12.19
gast==0.4.0
google-auth==2.48.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
groq
h11==0.16.0
h5py==3.15.1
idna==3.11
joblib==1.5.3
keras==3.12.0
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pillow==12.1.0
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
starlette==0.50.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
Werkzeug==3.1.5
wrapt==2.1.1